third_party/blink/web_tests/fast/dom/importNode-unsupported-node-type.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
description('importNode should throw informative errors for unsupported node types');

var doc = document.implementation.createHTMLDocument('');
shouldThrow("document.importNode(doc)")

var shadow = document.createElement('div').attachShadow({mode: 'open'});
shouldThrow("document.importNode(shadow)")
</script>
